fix api changes

+++ src/packet_sources/windows.rs
@@ -79,15 +79,15 @@ impl PacketSourceConf for WindowsConf {
         };
 
         if cfg!(debug_assertions) {
-            if result.0 <= 32 {
+            if result.0 as u32 <= 32 {
                 let err = windows::core::Error::from_win32();
                 log::warn!("Failed to start child process: {}", err);
             }
-        } else if result.0 == SE_ERR_ACCESSDENIED as isize {
+        } else if result.0 as u32 == SE_ERR_ACCESSDENIED {
             return Err(anyhow!(
                 "Failed to start the interception process as administrator."
             ));
-        } else if result.0 <= 32 {
+        } else if result.0 as u32 <= 32 {
             let err = windows::core::Error::from_win32();
             return Err(anyhow!("Failed to start the executable: {}", err));
         }
